runtime: printquoted: fix decoderune call
go.dev/cl/742305 had a stale signature for decoderune, which was updated
in go.dev/cl/725101 (converting an argument and return value from int to
uint).

Add appropriate conversions.

Reviewed-on: https://go-review.googlesource.com/c/go/+/748300
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>

+++ src/runtime/print.go
@@ -217,7 +217,7 @@ func printquoted(s string) {
 			continue
 		case runeError:
 			// Distinguish errors from a valid encoding of U+FFFD.
-			if _, j := decoderune(s, i); j == i+1 {
+			if _, j := decoderune(s, uint(i)); j == uint(i+1) {
 				gwrite(bytes(`\x`))
 				printhexopts(false, 2, uint64(s[i]))
 				continue
